Add custom configuration for newrelic in php
Currently it is not possible to add custom configuration in the file
newrelic.ini that lives in the newrelic_php_conf_dir directory, which
would make possible to enable features that are not enable by default
in New Relic (eg.: support to Guzzle, which requires an enable flag
in newrelic.ini file).

This change allows custom configurations to be added to the newrelic.ini
in a form of a key/value pairs.

Ref.:
https://docs.newrelic.com/docs/agents/php-agent/frameworks-libraries/guzzle